+++ wcomponents-examples/src/main/java/com/github/bordertech/wcomponents/examples/picker/TreePicker.java
@@ -14,6 +14,7 @@ import com.github.bordertech.wcomponents.WTextField;
 import com.github.bordertech.wcomponents.examples.common.ExplanatoryText;
 import com.github.bordertech.wcomponents.layout.ColumnLayout;
 import com.github.bordertech.wcomponents.layout.ListLayout;
+import com.github.bordertech.wcomponents.util.HtmlClassUtil;
 import com.github.bordertech.wcomponents.util.Util;
 import com.github.bordertech.wcomponents.validation.ValidatingAction;
 import java.text.SimpleDateFormat;
@@ -79,6 +80,7 @@ public class TreePicker extends WContainer {
 		mainPanel.setIdName("main_panel");
 		mainPanel.setLayout(new ColumnLayout(COL_WIDTH, COL_ALIGN, COL_HGAP, COL_VGAP));
 		mainPanel.setMargin(new Margin(COL_HGAP));
+		mainPanel.setHtmlClass(HtmlClassUtil.HtmlClassName.RESPOND);
 
 		buildUI();
 	}